Allows to not en/decrypt tuples on some commands
Commands can be flagged with SetNotEncryptDecryptTTS() to tells the
table AM to not encrypt or decrypt tuples from TTS when writing or
reading.

Flags are identified by the couple CurrentTransactionId /
CurrentCommandId and stored in a local to backend HTAB.

This is useful for VACUUM FULL/CLUSTER, in this case we do not desire
to decrypt / encrypt the tuples but just want to rewrite them without
touching them.