CREATE TABLE AS statement support
This commit introduces support of CREATE TABLE .. USING tcleam AS ..
statements.

CTAS is a special case that needs a transient AES key built right
before the statement is executed by standard_ProcessUtility(). To
handle this transient key the local HTAB has been reviewed: we do
not rely anymore on CurrentCommandId and CurrentTransactionID as key
because they are not set before standard_ProcessUtility(), so we
now rely on local transaction id (lxid). Relying only on lxid does
not guarantee unicity of HTAB entries built inside the same
transaction, but it does not matter because if current command
succeeds, its HTAB entry will be removed, then the HTAB key (lxid)
can be safely reused. If current command fails, the transaction is
aborted and the HTAB destroyed.